use input validation right

+++ src/main/java/de/jugda/registration/DeleteResource.java
@@ -6,7 +6,8 @@ import io.quarkus.qute.Template;
 import io.quarkus.qute.TemplateInstance;
 
 import javax.inject.Inject;
-import javax.validation.Valid;
+import javax.validation.ConstraintViolation;
+import javax.validation.Validator;
 import javax.ws.rs.BeanParam;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -16,6 +17,7 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
+import java.util.Set;
 
 /**
  * @author Niko Köbler, http://www.n-k.de, @dasniko
@@ -27,6 +29,8 @@ public class DeleteResource {
     @Inject
     DeleteService deleteService;
     @Inject
+    Validator validator;
+    @Inject
     Template delete;
     @Inject
     Template delete_thanks;
@@ -34,7 +38,9 @@ public class DeleteResource {
     @GET
     public TemplateInstance getDeleteForm(@QueryParam("eventId") String eventId, @QueryParam("id") String id) {
         if (id == null) {
-            return delete.data("eventId", eventId);
+            DeregistrationForm form = new DeregistrationForm();
+            form.setEventId(eventId);
+            return delete.data(form);
         } else {
             String name = deleteService.deleteFromUri(id);
             return delete_thanks.data("name", name);
@@ -43,9 +49,16 @@ public class DeleteResource {
 
     @POST
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
-    public TemplateInstance postDeleteForm(@Valid @BeanParam DeregistrationForm deregistrationForm) {
-        String name = deleteService.deleteFromUi(deregistrationForm);
-        return delete_thanks.data("name", name);
+    public TemplateInstance postDeleteForm(@BeanParam DeregistrationForm deregistrationForm) {
+        Set<ConstraintViolation<DeregistrationForm>> violations = validator.validate(deregistrationForm);
+        if (violations.isEmpty()) {
+            String name = deleteService.deleteFromUi(deregistrationForm);
+            return delete_thanks.data("name", name);
+        } else {
+            violations.forEach(cv ->
+                deregistrationForm.addValidationError(cv.getPropertyPath().toString(), cv.getMessage()));
+            return delete.data(deregistrationForm);
+        }
     }
 
     @DELETE
